full error resolve 06/Nov/2022

+++ app/Http/Controllers/Client/ClientProfileController.php
@@ -21,6 +21,18 @@ class ClientProfileController extends Controller
 
     public function updateProfile(Request $request)
     {
+        $validatedData = $request->validate([
+            'image' => 'image|mimes:jpg,jpeg,png,gif|max:2048',
+        ],
+        [
+            'image' => 'Please Choose Only jpg,jpeg,png,gif file.',
+        ]
+    
+    );
+
+        // echo "<pre>";
+        // print_r($request->all());
+        // die();
         $profileUpdate = User::find($request->id);
         $profileUpdate->name = $request->name;
         $profileUpdate->dob = $request->dob;
@@ -29,23 +41,37 @@ class ClientProfileController extends Controller
         // $profileUpdate->dob = $request->dob;
         $profileUpdate->address = $request->address;
 
+        // $profileUpdate->street = $request->street;
+        // $profileUpdate->city = $request->city;
+        // $profileUpdate->state = $request->state;
+        // $profileUpdate->country = $request->country;
+        // $profileUpdate->zip_code = $request->zip_code;
+
 
         if ($request->file('image') != NULL) {
 
             // Old Image Delete Code Start
             if ($profileUpdate->image != NULL) {
-                unlink('image/' . $profileUpdate->image);
+
+                if(file_exists('image/'. $profileUpdate->image)){
+                    if(file_exists('image/'. $profileUpdate->image)){
+                        unlink('image/'. $profileUpdate->image);
+                    }
+                   
+                }
+                
             }
             // Old Image Delete Code End
             $image = $request->file('image');
             $destinationPath = 'image';
             $uploadImage = date('YmdHis') . "." . $image->getClientOriginalExtension();
-            $image->move($destinationPath, $uploadImage);
+            // $image->move($destinationPath, $uploadImage);
+            $image->move(public_path($destinationPath), $uploadImage);
             $profileUpdate->image =  $uploadImage;
         }
         $profileUpdate->update();
 
-        return redirect()->back()->with('success', 'Profile Update Successfully Changed......!');
+        return redirect()->back()->with('success', 'Profile Updated Successfully......!');
 
     }
 
@@ -56,11 +82,16 @@ class ClientProfileController extends Controller
             'current_password' => ['required', new ClientOldPassword],
             'new_password' => ['required'],
             'new_confirm_password' => ['same:new_password'],
-        ]);
+        ],
+        [
+            'new_confirm_password' => 'Password does not match',
+        ]
+    
+    );
 
         User::find(auth()->user()->id)->update(['password' => Hash::make($request->new_password)]);
 
         // dd('Password change successfully.');
-        return redirect()->back()->with('password', 'Password Successfully Changed......!');
+        return redirect()->back()->with('password', 'Password Updated Successfully......!');
     }
 }
